feat: added checks for track and playlist access

+++ Sonar.UserTracksManagement/Sonar.UserTracksManagement.Application/Services/PlaylistApplicationService.cs
@@ -13,13 +13,15 @@ public class PlaylistApplicationService : IPlaylistApplicationService
 {
 
     private readonly IPlaylistService _playlistService;
+    private readonly ICheckAvailabilityService _availabilityService;
     private readonly IAuthorizationService _authorizationService;
     private readonly UserTracksManagementDatabaseContext _databaseContext;
-    public PlaylistApplicationService(IPlaylistService playlistService, IAuthorizationService authorizationService, UserTracksManagementDatabaseContext databaseContext)
+    public PlaylistApplicationService(IPlaylistService playlistService, IAuthorizationService authorizationService, UserTracksManagementDatabaseContext databaseContext, ICheckAvailabilityService availabilityService)
     {
         _playlistService = playlistService;
         _authorizationService = authorizationService;
         _databaseContext = databaseContext;
+        _availabilityService = availabilityService;
     }
     public async Task<Guid> CreateAsync(string token, string name)
     {
@@ -40,6 +42,10 @@ public class PlaylistApplicationService : IPlaylistApplicationService
             throw new NotFoundArgumentsException("Couldn't find playlist with given ID");
         if (track == null)
             throw new NotFoundArgumentsException("Couldn't find track with given ID");
+        if (!_availabilityService.CheckPlaylistAvailability(user.Id, playlist))
+            throw new PreconditionException("User doesn't have access to given playlist");
+        if (!_availabilityService.CheckTrackAvailability(user.Id, track))
+            throw new PreconditionException("User doesn't have access to given track");
         _playlistService.AddTrackToPlaylist(playlist, track);
     }
 
@@ -54,6 +60,10 @@ public class PlaylistApplicationService : IPlaylistApplicationService
             throw new NotFoundArgumentsException("Couldn't find playlist with given ID");
         if (track == null)
             throw new NotFoundArgumentsException("Couldn't find track with given ID");
+        if (!_availabilityService.CheckPlaylistAvailability(user.Id, playlist))
+            throw new PreconditionException("User doesn't have access to given playlist");
+        if (!_availabilityService.CheckTrackAvailability(user.Id, track))
+            throw new PreconditionException("User doesn't have access to given track");
         if (playlist.Tracks.All(item => item.Track.Id != trackId))
             throw new NotFoundArgumentsException("No track with given ID in the playlist");
         _playlistService.RemoveTrackFromPlaylist(playlist, track);
@@ -67,6 +77,8 @@ public class PlaylistApplicationService : IPlaylistApplicationService
         var playlist = await _databaseContext.Playlists.FirstOrDefaultAsync(item => item.Id.Equals(playlistId));
         if (playlist == null)
             throw new NotFoundArgumentsException("Couldn't find playlist with given ID");
+        if (!_availabilityService.CheckPlaylistAvailability(user.Id, playlist))
+            throw new PreconditionException("User doesn't have access to given playlist");
         return _playlistService.GetTracksFromPlaylist(playlist);
     }
